// src/types/generated/graphql.zonos-customer-graph.types.ts
export type Maybe<T> = T | null;
// ...
export type Carrier = {
  __typename?: 'Carrier';
  /** A unique identifier tied to a Carrier. */
  code: Scalars['String'];
  /** A list of country-specific restrictions a `Carrier` should be aware of (e.g., weight, dimensions, pricing). */
  countryConstraints: Maybe<Array<CountryConstraint>>;
  /** When this Carrier was created. */
  createdAt: Scalars['DateTime'];
  /** The user who created the Carrier. */
  createdBy: Scalars['ID'];
  /** The Carrier's API credentials. */
  credentials: Maybe<Array<CarrierCredential>>;
  /** A unique identifier for the Carrier. */
  id: Scalars['ID'];
  /** The humanly-memorable display name for the Carrier. */
  name: Scalars['String'];
  parties: Maybe<Array<Maybe<Party>>>;
  /** Provides a list of `ServiceLevel`s that are supported by the Carrier. */
  serviceLevels: Maybe<Array<ServiceLevel>>;
  /** When this Carrier was most recently updated. */
  updatedAt: Scalars['DateTime'];
  /** The user who most recently updated the Carrier. */
  updatedBy: Scalars['ID'];
};

/**
 * A `CarrierAccount` is an account number that is associated with a carrier for a given customer or group of customers.
 * Rates may vary across carrier accounts, depending on the customer and volume with the carrier.
 * Shipments cannot be created without a `carrierAccount`.
 */
export type CarrierAccount = {
  __typename?: 'CarrierAccount';
  /** A monetary amount specified by the merchant as a buffer for any additional fees. */
  additionalFee: Maybe<Scalars['Decimal']>;
  /** The `Carrier` associated with this CarrierAccount. */
  carrier: Carrier;
  /** When the CarrierAccount was created. */
  createdAt: Scalars['DateTime'];
  /** The user who created the CarrierAccount. */
  createdBy: Scalars['ID'];
  /** Where the credentials are stored. */
  credentialProvider: CredentialProvider;
  /** The `Carrier`'s API credentials. */
  credentials: Maybe<Array<Maybe<CarrierAccountCredential>>>;
  /** A unique identifier for the CarrierAccount. */
  id: Scalars['ID'];
  /** Specifies whether the CarrierAccount is in live or test mode. */
  mode: Mode;
  /** The `Organization` associated with the CarrierAccount. */
  organization: Scalars['ID'];
  /** Describes the `additionalFee` as a percentage. */
  shippingPercentIncrease: Maybe<Scalars['Decimal']>;
  /** When the CarrierAccount was most recently updated. */
  updatedAt: Scalars['DateTime'];
  /** The user who most recently updated the CarrierAccount. */
  updatedBy: Scalars['ID'];
};
// ...
export type Carton = {
  __typename?: 'Carton';
  /** The monetary amount insured for the `packingCarton` */
  amountInsured: Maybe<Scalars['Decimal']>;
  /** When this `Carton` was created */
  createdAt: Scalars['DateTime'];
  /** The user who created the `Carton` */
  createdBy: Scalars['ID'];
  /** The measurement units of the height, length and width */
  dimensionalUnit: DimensionalUnitCode;
  /** The numeric height of the `packingCarton` */
  height: Scalars['Decimal'];
  /** Carton ID, prefixed with `carton_` */
  id: Scalars['ID'];
  /** The items associated with the `packingCarton` */
  items: Maybe<Array<CartonItem>>;
  /** The numeric length of the `packingCarton` */
  length: Scalars['Decimal'];
  /** Whether the `Carton` is in test or live mode */
  mode: Mode;
  organization: Scalars['ID'];
  /** The packaging option associated with the `packingCarton` */
  packagingOption: Maybe<PackagingOption>;
  /** The id of the `root` that will own the `Carton`. */
  rootId: Scalars['ID'];
  /** Where the item is shipping from */
  shipFrom: Maybe<Location>;
  /** The order's destination (customer address) */
  shipTo: Maybe<Location>;
  /** The packaging style (box, polybag, letter, etc) */
  type: PackagingType;
  /** When this `Carton` was most recently updated. */
  updatedAt: Scalars['DateTime'];
  /** The user who most recently updated the `Carton`. */
  updatedBy: Scalars['ID'];
  /** The heaviness of the packingCarton, using a numeric value */
  weight: Scalars['Decimal'];
  /** The type of weight associated with the `packingCarton` */
  weightUnit: WeightUnitCode;
  /** The numeric width of the `packingCarton` */
  width: Scalars['Decimal'];
};
// ...
export type CurrencyFormat = {
  __typename?: 'CurrencyFormat';
  /** The currency code that applies to the `currencyFormat` */
  currencyCode: CurrencyCode;
  /** The character used to separate the whole units from the fractional units */
  decimalDelimiter: Scalars['String'];
  /** The number of decimal places to display for the currency */
  scale: Scalars['Int'];
  /** The characters used as a symbol of the currency */
  symbol: Scalars['String'];
  /** The location of where the symbol is displayed */
  symbolLocation: Maybe<CurrencySymbolLocation>;
  /** The character used as the thousands separate */
  thousandsDelimiter: Scalars['String'];
};
// ...
export type CustomsSpec = Node & {
  __typename?: 'CustomsSpec';
  /** Subtotal amounts of how the `Order` amount was calculated. */
  amountSubtotals: CustomsSpecAmountSubtotals;
  /** Objects including other details about the `CustomsSpec` */
  attributes: Maybe<Array<CustomsSpecAttribute>>;
  /** The `Carton` objects that are tied to this `CustomsSpec` */
  cartons: Maybe<Array<Carton>>;
  /** The type of clearance for this `CustomsSpec` */
  clearanceType: Maybe<ClearanceType>;
  /** The timestamp of when this `CustomsSpec` was created */
  createdAt: Scalars['DateTime'];
  /** The two letter currency code that the totals for this `CustomsSpec` will be represented in */
  currencyCode: CurrencyCode;
  /** The `CustomsItem` objects that are a result of generating this `CustomsSpec` */
  customsItems: Maybe<Array<CustomsItem>>;
  /** Declaration statement to be used for this `CustomsSpec` */
  declarationStatement: Maybe<Scalars['String']>;
  /** `CustomsSpec` ID prefixed with `customs_spec_` */
  id: Scalars['ID'];
  /** Applicable Incoterm for this `CustomsSpec` */
  incoterm: IncotermCode;
  /** Optional metadata key/value pairs */
  metadata: Maybe<Array<CustomsMetadata>>;
  /** The `Mode` this `CustomsSpec` was created in */
  mode: Mode;
  /** The `Organization` associated with the `CustomsSpec` */
  organization: Scalars['ID'];
  /** A list of parties associated with the `CustomsSpec` */
  parties: Maybe<Array<Party>>;
  /** The customs term for the end use of this export */
  reasonForExport: LandedCostEndUse;
  /** The service level that was used by the carrier for this `CustomsSpec` */
  serviceLevel: Maybe<Scalars['String']>;
  /** The tracking number provided by the carrier who is handling this shipment */
  trackingNumber: Maybe<Scalars['String']>;
  /** The timestamp of when this `CustomsSpec` was updated */
  updatedAt: Scalars['DateTime'];
};
// ...
export type CustomsSpecConnection = {
  __typename?: 'CustomsSpecConnection';
  edges: Maybe<Array<Maybe<CustomsSpecEdge>>>;
  pageInfo: Maybe<PageInfo>;
  totalCount: Maybe<Scalars['Int']>;
};
// ...
export type ExchangeRate = {
  __typename?: 'ExchangeRate';
  createdAt: Scalars['DateTime'];
  createdBy: Scalars['ID'];
  expiresAt: Scalars['DateTime'];
  id: Scalars['ID'];
  mode: Mode;
  rate: Scalars['Decimal'];
  sourceCurrencyCode: CurrencyCode;
  targetCurrencyCode: CurrencyCode;
  targetFormat: Maybe<CurrencyFormat>;
  type: ExchangeRateType;
};
// ...
export type ItemConnection = {
  __typename?: 'ItemConnection';
  /** Field edges */
  edges: Maybe<Array<Maybe<ItemEdge>>>;
  /** Field pageInfo */
  pageInfo: Maybe<PageInfo>;
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Returns a `Carrier` resource by ID. */
  carrier: Maybe<Carrier>;
  /** Returns a `CarrierAccount` by ID */
  carrierAccount: Maybe<CarrierAccount>;
  /** Returns all the information tied to the `Carrier` that a specified organization uses. */
  carrierAccounts: Maybe<Array<Maybe<CarrierAccount>>>;
  /** Returns  a `Carrier` based on a `string` code. */
  carrierByCode: Maybe<Carrier>;
  /** Returns a list of `Carrier` resources by ID. */
  carriers: Maybe<Array<Maybe<Carrier>>>;
  /** Returns a `Carton` resource by `ID` */
  carton: Maybe<Carton>;
  /** Returns a list of `Carton` objects that apply to a `root` */
  cartons: Maybe<Array<Maybe<Carton>>>;
  currencyFormat: Maybe<CurrencyFormat>;
  /** Retrieve a `CustomsSpec` by ID */
  customsSpec: Maybe<CustomsSpec>;
  /** Retrieve a list of `CustomsSpec` objects */
  customsSpecs: Maybe<CustomsSpecConnection>;
  exchangeRate: Maybe<ExchangeRate>;
  /** Returns a `FulfillmentCenter` object by ID. */
  fulfillmentCenter: Maybe<FulfillmentCenter>;
  /** Returns a list of `FulfillmentCenter` objects for an `Organization` */
  fulfillmentCenters: Maybe<Array<Maybe<FulfillmentCenter>>>;
  getPerson: Maybe<Person>;
  /** Returns an `Item` resource by ID. */
  item: Maybe<Item>;
  itemRestrictionResult: Maybe<ItemRestrictionResult>;
  /** Returns a list of `Item` resources. */
  items: Maybe<ItemConnection>;
  /** Returns a `LandedCost` resource by `ID` */
  landedCost: Maybe<LandedCost>;
  /** Returns a list of `LandedCost` resources */
  landedCosts: Maybe<LandedCostConnection>;
  location: Maybe<Location>;
  order: Maybe<Order>;
  orders: Maybe<OrderConnection>;
  /** Returns a `PackagingOption` resource by `ID` */
  packagingOption: Maybe<PackagingOption>;
  /** Returns a list of `PackagingOption` objects that apply to an `organization` */
  packagingOptions: Maybe<Array<Maybe<PackagingOption>>>;
  parties: Maybe<Array<Maybe<Party>>>;
  partiesFindAllById: Maybe<Array<Maybe<Party>>>;
  party: Maybe<Party>;
  /** Return a PDDP submission by an ID. */
  pddpSubmission: Maybe<PddpSubmission>;
  /** Return a PDDP submission log by an ID. */
  pddpSubmissionLog: Maybe<PddpSubmissionLog>;
  /** Returns a paginated list of PDDP submissions logs. */
  pddpSubmissionLogs: Maybe<PddpSubmissionLogConnection>;
  /** Returns a paginated list of PDDP submissions. */
  pddpSubmissions: Maybe<PddpSubmissionConnection>;
  /** Returns a `ServiceLevel` resource by ID. */
  serviceLevel: Maybe<ServiceLevel>;
  /** Returns a `ServiceLevelArea` resource by ID. */
  serviceLevelArea: Maybe<ServiceLevelArea>;
  /** Retreive the ServiceLevelAreas for a giving ServiceLevel optionally filtering by a destination location */
  serviceLevelAreas: Maybe<Array<Maybe<ServiceLevelArea>>>;
  /** Returns a `ServiceLevelRateChart` by ID. */
  serviceLevelRateChart: Maybe<ServiceLevelRateChart>;
  /** Returns a list of `ServiceLevelRateChart` objects that apply to a `ServiceLevel` */
  serviceLevelRateCharts: Maybe<Array<Maybe<ServiceLevelRateChart>>>;
  /** Returns a list of `ServiceLevel` resources for an organization. */
  serviceLevels: Maybe<Array<Maybe<ServiceLevel>>>;
  shipment: Maybe<Shipment>;
  /** Returns a `ShipmentRating` resource by ID. */
  shipmentRating: Maybe<ShipmentRating>;
  /** Returns a list  of `ShipmentRating` resources with the given filters. */
  shipmentRatings: Maybe<Array<Maybe<ShipmentRating>>>;
  shipments: Maybe<ShipmentConnection>;
  /** Returns a `ShippingProfile` by ID. */
  shippingProfile: Maybe<ShippingProfile>;
  /** Returns a `ShippingProfileRateChart` by ID */
  shippingProfileRateChart: Maybe<ShippingProfileRateChart>;
  /** Returns a list of `ShippingProfileRateChart` objects that apply to a `ShippingProfile` */
  shippingProfileRateCharts: Maybe<Array<Maybe<ShippingProfileRateChart>>>;
  /** Returns a list of `ShippingProfile` objects that apply to an `Organization`. */
  shippingProfiles: Maybe<Array<Maybe<ShippingProfile>>>;
  /** Returns a Shipping Store Settings resource by store id. */
  shippingStoreSettings: Maybe<ShippingStoreSettings>;
  /** Returns a `ShippingZone` resource by ID. */
  shippingZone: Maybe<ShippingZone>;
  /** Returns a list of `ShippingZone` resources for an organization. */
  shippingZones: Maybe<Array<Maybe<ShippingZone>>>;
  taxIds: Maybe<Array<Maybe<TaxId>>>;
};
// ...
export type CurrencyFormatQuery = { __typename?: 'Query' } & {
  currencyFormat: Maybe<
    { __typename?: 'CurrencyFormat' } & Pick<
      CurrencyFormat,
      | 'decimalDelimiter'
      | 'thousandsDelimiter'
      | 'scale'
      | 'symbol'
      | 'symbolLocation'
    >
  >;
};